Finished delta message, fixed Deltas being measured in ints

+++ WindowSystem.cpp
@@ -30,7 +30,7 @@ WindowSystem::WindowSystem(int gwidth, int gheight, char* gtitle){
 	initializeWindow();
 }
 
-void WindowSystem::update(int delta){
+void WindowSystem::update(float delta){
     glViewport(0,0,width,height);
     glClearColor(0.0f, 0.0f, 1.0f, 0.5f);
